the laravel course the test

+++ modules/Cyaxaress/Course/Http/Controllers/CourseController.php
@@ -51,9 +51,11 @@ class CourseController extends Controller
     public function update($id, CourseRequest $request, CourseRepo $courseRepo)
     {
         $course = $courseRepo->findbyId($id);
+        $this->authorize('edit', $course);
         if ($request->hasFile('image')) {
             $request->request->add(['banner_id' => MediaFileService::upload($request->file('image'))->id]);
-            $course->banner->delete();
+            if ($course->banner)
+                $course->banner->delete();
         } else {
             $request->request->add(['banner_id' => $course->banner_id]);
         }
@@ -66,6 +68,7 @@ class CourseController extends Controller
     public function destroy($id, CourseRepo $courseRepo)
     {
         $course = $courseRepo->findbyId($id);
+        $this->authorize('delete', $course);
         if ($course->banner) {
             $course->banner->delete();
         }
@@ -75,6 +78,7 @@ class CourseController extends Controller
 
     public function accept($id, CourseRepo $courseRepo)
     {
+        $this->authorize('change_confirmation_status', Course::class);
         if ($courseRepo->updateConfimaitonStatus($id, Course::CONFIRMATION_STATUS_ACCEPTED)) {
             return AjaxResponses::SuccessResponse();
         }
@@ -83,6 +87,7 @@ class CourseController extends Controller
 
     public function reject($id, CourseRepo $courseRepo)
     {
+        $this->authorize('change_confirmation_status', Course::class);
         if ($courseRepo->updateConfimaitonStatus($id, Course::CONFIRMATION_STATUS_REJECTED)) {
             return AjaxResponses::SuccessResponse();
         }
@@ -91,6 +96,7 @@ class CourseController extends Controller
 
     public function lock($id, CourseRepo $courseRepo)
     {
+        $this->authorize('change_confirmation_status', Course::class);
         if ($courseRepo->updatStatus($id, Course::STATUS_LOCKED)) {
             return AjaxResponses::SuccessResponse();
         }
